feat: Github 데이터 저장하고 렌더링하는 함수 추가

+++ src/main/java/com/swcoaching/example1/domain/github/GithubData.java
@@ -17,19 +17,19 @@ public class GithubData extends BaseTimeEntity {
     private Long id;
     private String githubUsername;
     @Column(name = "date_year")
-    private int year;
-    private int starCount;
-    private int commitCount;
-    private int prCount;
-    private int issueCount;
-    private int followerCount;
-    private int followingCount;
-    private int repoCount;
+    private Long year;
+    private Long starCount;
+    private Long commitCount;
+    private Long prCount;
+    private Long issueCount;
+    private Long followerCount;
+    private Long followingCount;
+    private Long repoCount;
     private Long userId;
 
     @Builder
-    public GithubData(Long id, String githubUsername, int year, int starCount, int commitCount, int prCount,
-                      int issueCount, int followerCount, int followingCount, int repoCount) {
+    public GithubData(Long id, String githubUsername, Long year, Long starCount, Long commitCount, Long prCount,
+                      Long issueCount, Long followerCount, Long followingCount, Long repoCount) {
         this.id = id;
         this.githubUsername = githubUsername;
         this.year = year;
